[clang][dataflow] Fix comments, names, spacing, ordering.
Also move RecordInitListHelper to ASTOps.
